+++ src/pages/[...slug].astro
@@ -2,7 +2,7 @@
 import { getCollection, render } from 'astro:content';
 
 import CodeBlock from '@/components/article/CodeBlock.astro';
-import CustomLink from '@/components/article/CustomLink';
+import { CustomLink } from '@/components/article/CustomLink';
 import CustomTable from '@/components/article/CustomTable.astro';
 import FragmentTitle from '@/components/article/FragmentTitle.astro';
 import ResponsiveImage from '@/components/article/ResponsiveImage.astro';
